Add: RemoteStreamConn and GracefulCloseWebSocket
